we got it working!!!
Needs testing, but currently Planets is alive and well!  Studied the finer details of the this keyword and how constructors work, and with a little guidance from a friendly programmer on r/javascript (u/albedoa) it's finally performing as intended (at least, on a first pass structural level). Next comes rigorous testing of each Method of the class to ensure it's outputting as desired.